backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
SQLAlchemy>=2.0
python-dotenv>=1.0
python-jose[cryptography]>=3.3
passlib==1.7.4
bcrypt==4.0.1
pydantic>=2.5
email-validator>=2.1
psycopg2-binary>=2.9